Add support for blpop and brpop
The header for commands is no longer written when the command is
constructed since the number of arguments is not known until command
execution time.  A custom execute method is added that works for both
the blpop and brpop commands and it takes a list of strings representing
the keys for the lists to perform the operation against and an integer
representing the timeout value for the operation.

These commands can return a nil multi-bulk reply which is indicated by a
-1 count in the multi-bulk reply.  Instead of throwing an exception we
now return false and unlink the MultiBulkEnumerator from the list of
outstanding replies.